attention kernel output f32 -> f16

+++ tests/test_attention.py
@@ -36,7 +36,7 @@ def test_attention_correctness(device, B, H, N_CTX, D_HEAD):
     
     # 3. Softmax along key-dimension and multiply with V
     p = torch.softmax(scores, dim=-1)
-    o_ref = torch.matmul(p, v)
+    o_ref = torch.matmul(p, v).to(torch.float16)
 
     # Higher tolerances because Triton online softmax (FP32 accumulations) and 
     # eager torch FP16 materializations have slight arithmetic rounding variations.
@@ -72,7 +72,7 @@ def test_attention_non_power_of_two_shapes(device, B, H, N_CTX, D_HEAD):
     mask = torch.triu(torch.full((N_CTX, N_CTX), float('-inf'), device=device), diagonal=1)
     scores = scores + mask[None, None, :, :]
     p = torch.softmax(scores, dim=-1)
-    o_ref = torch.matmul(p, v)
+    o_ref = torch.matmul(p, v).to(torch.float16)
 
     torch.testing.assert_close(
         o_triton,
